requirements.txt
asgiref==3.2.3
backcall==0.1.0
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
Django==3.0.2
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-robots==4.0
Fabric==1.14.0
gunicorn==20.0.4
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
paramiko==2.7.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
PyNaCl==1.3.0
pytils==0.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
